+++ addon/components/flexberry-simpleolv.js
@@ -2622,7 +2622,8 @@ export default folv.extend(
         buttonName: '...', // Button displayed name.
         buttonAction: '...', // Action that is called from controller on this button click (it has to be registered at component).
         buttonClasses: '...', // Css classes for button.
-        buttonTitle: '...' // Button title.
+        buttonTitle: '...', // Button title.
+        disabled: true, // The state of the button is disabled if `true` or enabled if `false`.
       }
       ```