+++ tests/dummy/app/models/ember-flexberry-dummy-localized-suggestion-type.js
@@ -22,6 +22,11 @@ var Model = Projection.Model.extend({
       presence: {
         message: 'Name is required'
       }
+    },
+    localization: {
+      presence: {
+        message: 'Localization is required'
+      }
     }
   }
 });